+++ frontend/src/App.jsx
@@ -5,10 +5,12 @@ import MemberCreatePage from "./pages/MemberCreatePage";
 import AssetPage from "./pages/AssetPage";
 import AssetCreatePage from "./pages/AssetCreatePage";
 import AssetItemPage from "./pages/AssetItemPage";
+import AssetItemCreatePage from "./pages/AssetItemCreatePage";
 import CategoryPage from "./pages/CategoryPage";
 import HomePage from "./pages/HomePage";
 import LoanPage from "./pages/LoanPage";
 import ReservationPage from "./pages/ReservationPage";
+import ReservationCreatePage from "./pages/ReservationCreatePage";
 import Notfound from "./pages/Notfound";
 
 function App() {
@@ -21,10 +23,12 @@ function App() {
         <Route path="/assets" element={<AssetPage />} />
         <Route path="/assets/new" element={<AssetCreatePage />} />
         <Route path="/asset-items" element={<AssetItemPage />} />
+        <Route path="/asset-items/new" element={<AssetItemCreatePage />} />
         <Route path="/categories" element={<CategoryPage />} />
 
         <Route path="/loans" element={<LoanPage />} />
         <Route path="/reservations" element={<ReservationPage />} />
+        <Route path="/reservations/new" element={<ReservationCreatePage />} />
         <Route path="*" element={<Notfound />} />
       </Routes>
     </>